feat(cubestore): cleanup local copies of files removed remotely
The goal is to free up disks on the worker machines. This process runs
every 10 minutes.

This is done via a background process that periodically lists files on
remotes and removes any extra files we have locally.

The cleanup only affects the direct siblings (i.e. the `.parquet` files)
and not subdirectories. E.g. the exceptions to the cleanup are
`metastore-*` and `node-heart-beats`. The underlying subsystems manage
their local copies directly.

Uploads are now prepared in a separate subdirectory of `.cubestore` to
avoid accidental removals by the cleanup process. The `RemoteFS` APIs
are updated accordingly.