QueryBuilder Improvements - types & safety
Updates TypeScript types to be more accurate (and fix some mistakes),
adding PropTypes for non-typescript safety.  Refactoring updateVizState
to be more consistent and easier to understand, updating orderMember
handling to be less problematic and use a single source of truth.
Convert to a fully uncontrolled component to follow react best practices
and prevent unexpected behavior.