chore(cubestore): refactor query planning
We now use datafusion interface in a more direct manner.
Planning is expressed as a series of transformations on logical and
physical plans.

This change temporarily removes the optimization that sends the
single-node queries fully to the worker.
However, the change should give a good start for more sophisticated
optimizations in the future.