feat(cubestore): top-k query planning and execution
This allows to more efficiently execute queries like:

```
SELECT group_key, SUM(agg) FROM … GROUP BY 1 ORDER BY 2 LIMIT 10
```

Previously, we would:
    1. pre-aggregate results on the workers,
    2. send pre-aggregated results from each worker to the router,
    3. do the final aggregation on the router for the full data,
    4. sort full results,
    5. return top 10 rows.

When the dataset is large and the limit is small, we clearly do a
lot more work than needed, both on the workers and on the router.

After this change, we:
    1. pre-aggregate and sort on the workers,
    2. send results to the router,
    3. process only the required number of rows from each worker to get
       the 10 result rows.

After this change, we do much less work on the router.
More improvements are on the way:
  - stream only the required number of rows from each worker,
  - allow to replace the 'sort and pre-aggregate' step on the workers
    with table scan from an index if possible.