fix(cubestore): make top-k scan less batches
First scan the nodes required to make progress on the top candidate.

Improves performance when the top values are missing on some of the nodes.
We need to scan these nodes until the end, but we now do no unnecessary scans of
other nodes that already produced the top values.