fix(cubestore): send at most one request per worker
We would previously send one request per partition.
When many partitions are on the same worker, more than one request to
the same worker is clearly redundant.

This is especially important for top-k queries where each partition
of `ClusterSendExec` adds overhead.

Some planning tests now depend on cluster configuration. This leads to
failures on 'cluster' or 'in-process' SQL tests. To avoid failures,
temporarily disabled the test in cluster configuration.