feat: Execute refreshKeys in externalDb (only for every) (#3061)
Internal:
First, there is an internal change, that we started to pass query options as 3rd parameter in Query tuple.

Motivation
We are going to execute refresh keys (only for every, just now) in an external database. In the feature, we are planning to improve support for every via changes in the cube store (support more useful SQL things).