fix(cubestore): speed up HLL++ merges, up to 180x in some cases
Add a buffer that delays write into VLE-encoded array for Sparse HLLs.
This produces 45x improvement for merging 10k single-value HLLs.
This adds up to 180x improvement with the previous commit.

The buffer is a `BTreeSet` instead of vector in the original
implementation. This allows to keep the buffer smaller and do less
compactions in some cases, e.g. when merging HLLs with many common
elements. It does use more memory and CPU when merging HLLs with
many different hashes.