chore(cubestore): change non-repartitioned chunk assignment on queries
We used to assign all non-repartitioned chunks to leaf partitions.
Now we report them alongside parent partitions and skip main table of
parent partitions if it is not active.

We need this for proper functioning of partitioned indices.
Distribution of multi-partitions relies on having the parent
partitions to properly issue filters on chunks.

The current strategy has other problems:
- Data in chunks for partitions with >1 child gets duplicated
  in some cases. This leads to wrong results on some queries.
- We may do extra downloads for chunks of parent partition on
  workers handling child queries. This latency can be avoided
  as data is likely already on the worker for the parent
  partition.