Reverse proxy documentation
I've been working on deploying an httos cubejs instance using `docker-compose` over the past few days. Although it was good enough for me to finally make it, as an absolute Docker Compose beginner I ran into some issues that might have been obvious to a more experienced dev, but they weren't for me. Although it was a cool learning experience, I think updating the documentation to be more accurate and work out of the box will make things easier for those ahead.

The changes are:
- Specify that the `cube.conf` file goes inside a `nginx` directory, otherwise nothing will be mounted in the corresponding
- Don't mount the `nginx` directly on top of `etc/nginx`, do it on `etc/nginx/conf.d` instead, otherwise it will overwrite existing nginx configuration, and the webserver will fail to start.
- In the nginx `proxy_pass` configuration, change the `localhost` reference to a `cube` reference; in Docker Compose it appears that each service has it's own separate `localhost`, and that configuration in Linux would cause nginx to fail with a 502 error. Referencing the `cube` service instead works as expected.
- Add some extra helpful lines, and a link to a good resource I found and followed to set up SSL with LetsEncrypt.
- Lastly, I changed the `webserver` service name to `nginx`. It's a small change and shouldn't be too important one way or another, but in this case the linked resource assumes the nginx service is named `nginx`, so why not