postgres driver add enum support
The user defined enum types were being dropped in the Postgres driver.
As per the postgres documentation these are stored internally as four
bytes on disk however the translations occur at query time to the label
from pg_enum which specifies:

"The length of an enum value's textual label is limited by the NAMEDATALEN
setting compiled into PostgreSQL; in standard builds this means at most
63 bytes." [1]

For the purposes of cubejs it should be acceptable to coerce into a
string.

[1] https://www.postgresql.org/docs/9.1/datatype-enum.html