fix(@cubejs-client/dx): reverse dependency order with `@cubejs-client/core`
Prior to this change, `@cubejs-client/dx` was a dependency of `@cubejs-client/core`. This approach caused significant problems with new projects generated with the Angular CLI, as it sets `skipLibCheck: false` in the `tsconfig.json` of newly-generated projects.

This commit reverses the dependency order; now `@cubejs-client/dx` wraps `@cubejs-client/core`, and exports "enhanced" types through its' `index.d.ts`. However, these types are effectively a copy-paste from `@cubejs-client/core` with slight modifications made so that it looks up CLI generated types if they exist. This is not a great approach, so we'll need to think of something more reliable long-term - but it can wait till after `@cubejs-client/core` is converted to TypeScript.