feat(cubestore): Repartition single chunks instead of partition as a whole to speed up ingestion of big tables (#4125)
In the case of ingestion of 200M+ row tables, the biggest bottleneck is the very first repartitioning. It requires the single node to download all the chunk files and scan through them. The solution is to repartition every single chunk separately and schedule nodes by chunk id in a round-robin fashion except for in-memory chunks where partitioning should be done on the partition owner node.