feat: persistence access mode
Using the default storage class of the AWS EKS cluster (gp2), the default
it is not possible to use the EBS volume for the persistent storage of the
Kubernetes cluster for the persistent volume claim

It should be possible to configure this via values.

The default value for the workersCount must be reduced to 1, because the
ReadWriteOnce volume mode do not support multiple mounts.