Make dimensions and measures nullable fields.
Fixes: https://github.com/cube-js/cube.js/issues/4399

Please see the discussion on this ticket on why this breaks things.

Currently the actual cubeJS schema does not have nullable types, and assuming all dimensions and measures are non-Nullable is not necessarily correct and blocking the use of the graphQL api entirely for cubes that can return null for dimensions or measures.

I think the ideal solution is to implement nullable types in the actual cubeJS schema so they can be translated 100% correct to a graphQL schema. However that is probably an overall bigger change.

For now, allowing graphQL Member fields to be nullable generates a more useable schema than the current nonNull configuration. (neither are 100% accurate though).