Add iam_role arn to Redshift Unload (#4509)
* Add iam_role arn to Redshift Unload

This makes the keys `keyId` and `secretKey` optional when `unloadArn` is set (with env: `CUBEJS_DB_REDSHIFT_UNLOAD_ARN`).  `secretKey` and `keyId` are still required when `unloadArn` is absent since Redshift needs to use credentials to perform `UNLOAD`.  If the redshift `iam_role` is used, keys should not be required for aws api operations if the proper task-roles are defined.

This also adds the `CUBEJS_DB_REDSHIFT_UNLOAD_ARN` environment variable, referenced in `env.ts` by `redshiftBucketUnloadArn`

fixes #4505

* Add documentation for Redshift Unload iam_role

Refs #4505

* Add tests

And fix the bug that was found by the test.

* Update ENV Naming

Use more consistent naming strategy for the Unload ARN Env variable.

* Formatting

Format the ternary operation to be more readable.

* Document the unloadARN Parameter

Add documentation and a reference link to the AWS documentation for the ARN role that needs to write the unload data.