feat(cubestore): Queue - optimize list/pending operations (#7998)
This PR separates a payload field (can be a large string) from `QueueItem` to a separate table called `QueueItemPayload`. Why? Because we have commands that iterate over the whole data. For example `QUEUE LIST` or `QUEUE TO_CANCEL` or `QUEUE PENDING`. It's super expensive to read and deserialize the payload from RocksDB.

For backward compatibility, I've bumped the version of QueueItem from `2` to `3`. For forward, and backward migration, Cache Store will truncate this table.

```
in_process::queue_list_bench
                        time:   [10.684 ms 10.769 ms 10.858 ms]
                        change: [-75.723% -74.952% -74.200%] (p = 0.00 < 0.05)
                        Performance has improved.
```