Update rest-api.mdx
Seems like this doc has not been updated since we released views. I updated all of the commentary to include mention of views based on my own testing.

Additionally, I added a note that anything with public: false is not returned in the results. I think this is weird behavior because in my tests I see a parameter for isVisible and a parameter for public. The existence of those parameters makes me think it should return private cubes and views so that sometimes public can say no, but since that's not what's happening, and since I don't know the difference between isVisible and public, I didn't add them to this PR, I just noted that public: false will make them omitted.