feat: Support complex join conditions for grouped joins (#9157)
Add support for complex join conditions for grouped joins

DataFusion plans non-trivial joins (ones that are not `l.column = r.column`) as `Filter(CrossJoin(...))`
To support ungrouped-grouped joins with queries like this SQL API needs to rewrite logical plan like that to `WrappedSelect` with join inside. To do that it need to distinguish between plan coming from regular `JOIN` and actual `CROSS JOIN` with `WHERE` on top. This is done with new `JoinCheckStage`: it starts on `Filter(CrossJoin(wrapper, wrapper))`, traverses all `AND`s in filter condition, checks that "leaves" in condition are comparing two join sides, and pulls up that fact. After that regular join rewrite can start on checked condition.

Supporting changes:

* Allow grouped join sides to have different in_projection flag

* Allow non-push_to_cube WrappedSelect in grouped subquery position in join

* Make zero members wrapper more expensive than filter member

* Replace alias to cube during wrapper pull up

* Wrap is_null expressions in parens, to avoid operator precedence issues
Expression like `(foo IS NOT NULL = bar IS NOT NULL)`` would try to compare `foo IS NOT NULL` with `bar`, not with `bar IS NOT NULL`