chore(cubestore): DF upgrade: Disable ident normalization, remove lowercase normalization in some lookups
This is to get ourselves back in line with old pre-DF-upgrade
behavior.  Maybe, instead, we should force Cube to quote literals in
its queries, but suppose we did that: We're working with generated
queries.  Normalization to lowercase would mean that any unquoted
identifiers that have uppercase characters would be a certain bug.

This avoids one factor that would require Cube changes and a Cube
upgrade in order to use the Cube.