refactor(query-orchestrator): LocalQueue driver - correct handling for pending/active (#9706)
I did a refactor for the `LocalQueue` driver. Previously, while the `LocalQueue` driver processed the queue item, it stored it in the pending and active states. Both states at the same time. This approach is not aligned with any queue servers. 

Cube Store has a strict obligation that a queue item can be in only one state at a time. Let's align it to simplify upcoming changes in https://github.com/cube-js/cube/pull/9705.