fix(cubestore): Make projection_above_limit optimization behave deter… (#9766)
* fix(cubestore): Make projection_above_limit optimization behave deterministically

This orders columns in plan generation based on the order they're seen
instead of using hash table ordering.  Note that this affects internal
plan nodes and does not change the output of any correctly-running
queries.  This has the effect of making query behavior deterministic
and reproducible when investigating other bugs in query evaluation.

* chore(cubestore): Make projection_above_limit::ColumnRecorder use IndexSet