fix(clickhouse): avoid placeholder conflicts by using explicit parameter tokens
The current implementation of the ClickHouse driver uses the `?` symbol as a placeholder. The `sqlstring.format` function replaces all `?` characters positionally with parameters. This only works when the built query does not contain `?` for any other purpose; it breaks as soon as a literal question mark is used in the SQL.

Question marks commonly appear in SQL literals, for example in regular expressions used in cube fields:

```
group: {
  title: `group`,
  type: `string`,
  sql: `extract(${CUBE}."group", '(.*?)(?:-?\\d{2})')`
}
```

There is also no way to escape such occurrences of the `?` symbol for `sqlstring.format`, so these conflicts cannot be avoided with the existing approach.

This change replaces the placeholder symbol `?` with placeholder tokens of the form `___ClickHouseParam_(\d+)___` in the driver and the clickhouse schema compiler. The remaining logic and escaping of sqlstring (using `sqlstring.escape`) stays intact.

Signed-off-by: W48B1T <125169428+W48B1T@users.noreply.github.com>